requirements.txt
fastapi[standard]==0.115.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.6.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
aiofiles==24.1.0
loguru==0.7.2
httpx==0.28.0
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
uvicorn[standard]==0.32.0 #server
